Width field of %c conversion is dynamically checked and must be 0 or 1.
The function end_of_input was wrong in case of a buffer for which both
beginning_of_input and end_of_input conditions were true (e.g. in case
of an empty string or an empty file).

Better error reporting for characters out of the given range.

Reminiscent of the printf semantics, %! matches the end of input
condition.