Prefer forward slashes to backward slashes on NT as it's more portable.
Nowdays all the Windows operating systems accept both, more over it's
being passed to our toolchain as a relative path (not UNC), so I see
no harm if we want to be a little bit more portable. The only reason
to not doing so is conventiob and IDEs which might parse Makefile.nt
and not recognise it.

Proposed by Adrien Nader.

git-svn-id: http://caml.inria.fr/svn/ocaml/trunk@13176 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02